requirements.txt
Django==4.1.2
djangorestframework==3.14.0
requests==2.28.1